Refacto FightManager, add Fight and CombatAction files
