+++ Assets/Scenes/MainScene.unity
@@ -2918,7 +2918,6 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   EnemyTeamPanel: {fileID: 545817102}
-  EnemyTeam: {fileID: 0}
 --- !u!114 &1700758566
 MonoBehaviour:
   m_ObjectHideFlags: 0
