+++ Assets/Scripts/Fight.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 664e499a16fab354e80f768ab10305b9
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
