+++ Assets/Scripts/Fight/CombatAction.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 823d1458422230a49824b9c1f8835d9f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
