+++ Assets/Scripts/Fight/Fight.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dcaf95768135f0c4b85df2ae9cc22a51
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
