+++ Assets/Scripts/Managers/Prompt.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ca95a73909f22fb409bd868eec1a99c3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
